Refactor sendRoundMessage to Session
